refactor: Extract components fetching from LightCube
